requirements.txt
Django==2.0.2
djangorestframework==3.7.7
pytz==2018.3
Pillow==5.0.0
xlwt==1.3.0
celery==4.1.0
django-model-utils==3.1.1

#psycopg2==2.7.3.2